feat: scale XWAU mesh pipeline for extended craft state

+++ include/aeron/scene/gltf_mesh.h
@@ -165,7 +165,7 @@ typedef struct AeronGltfModel {
     /* Merged geometry — one buffer per ship. The index buffer contains
      * stable opaque, alpha-mask, and alpha-blend ranges in that order. */
     AeronGltfVertex *vertices;     uint32_t vertex_count;
-    uint16_t         *indices;      uint32_t index_count;
+    uint32_t         *indices;      uint32_t index_count;
     uint32_t          opaque_index_count;
     uint32_t          mask_index_offset;
     uint32_t          mask_index_count;
